# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: GradeEscolar — Sistema Completo de Gestão de Horários Escolares
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /professores:
    get:
      operationId: listProfessores
      tags: [professores]
      summary: Lista todos os professores
      responses:
        "200":
          description: Lista de professores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Professor"
    post:
      operationId: createProfessor
      tags: [professores]
      summary: Cria um professor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfessorInput"
      responses:
        "201":
          description: Professor criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professor"
  /professores/{id}:
    get:
      operationId: getProfessor
      tags: [professores]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Professor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professor"
        "404":
          description: Não encontrado
    patch:
      operationId: updateProfessor
      tags: [professores]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfessorUpdate"
      responses:
        "200":
          description: Atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Professor"
    delete:
      operationId: deleteProfessor
      tags: [professores]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removido
  /professores/{id}/carga:
    get:
      operationId: getProfessorCarga
      tags: [professores]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Carga horária
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessorCarga"

  # ── DISCIPLINAS ───────────────────────────────────────────────────────────────
...
  /salas:
    get:
      operationId: listSalas
      tags: [salas]
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sala"
    post:
      operationId: createSala
      tags: [salas]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalaInput"
      responses:
        "201":
          description: Criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sala"
  /salas/{id}:
    patch:
      operationId: updateSala
      tags: [salas]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalaUpdate"
      responses:
        "200":
          description: Atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sala"
    delete:
      operationId: deleteSala
      tags: [salas]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removida

  # ── DISPONIBILIDADE ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Professor:
      type: object
      required: [id, nome, email, ativo, createdAt]
      properties:
        id: { type: integer }
        nome: { type: string }
        email: { type: string }
        telefone: { type: ["string", "null"] }
        ativo: { type: boolean }
        disciplinaIds:
          type: array
          items: { type: integer }
        createdAt: { type: string }
    ProfessorInput:
      type: object
      required: [nome, email]
      properties:
        nome: { type: string, minLength: 2 }
        email: { type: string }
        telefone: { type: string }
        disciplinaIds:
          type: array
          items: { type: integer }
    ProfessorUpdate:
      type: object
      properties:
        nome: { type: string }
        email: { type: string }
        telefone: { type: string }
        ativo: { type: boolean }
        disciplinaIds:
          type: array
          items: { type: integer }
    ProfessorCarga:
      type: object
      required: [professorId, totalAulas, porDia]
      properties:
        professorId: { type: integer }
        totalAulas: { type: integer }
        porDia:
          type: object
          additionalProperties: { type: integer }
...
    Sala:
      type: object
      required: [id, nome, tipo, capacidade, ativa]
      properties:
        id: { type: integer }
        nome: { type: string }
        tipo: { type: string, enum: [sala_aula, laboratorio, quadra, informatica, auditorio, biblioteca, sala_arte, outro] }
        capacidade: { type: integer }
        ativa: { type: boolean }
        observacoes: { type: ["string", "null"] }
    SalaInput:
      type: object
      required: [nome]
      properties:
        nome: { type: string, minLength: 1 }
        tipo: { type: string, enum: [sala_aula, laboratorio, quadra, informatica, auditorio, biblioteca, sala_arte, outro] }
        capacidade: { type: integer }
        observacoes: { type: string }
    SalaUpdate:
      type: object
      properties:
        nome: { type: string }
        tipo: { type: string }
        capacidade: { type: integer }
        ativa: { type: boolean }
        observacoes: { type: string }
...